Edit the items list of a order to the list of the PATCH request body

+++ Models/Item.py
@@ -1,8 +1,9 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 class Item(BaseModel):
     item_id: str
     name: str
     description: str | None = None
-    price: float
+    price: Decimal
     quantity: int
